untangle spack.config / spack.spec
