+++ lib/spack/spack/util/cpus.py
@@ -6,6 +6,7 @@
 import multiprocessing
 import os
 
+
 def cpus_available():
     """
     Returns the number of CPUs available for the current process, or the number
